% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_edgeevidence}
\alias{plot_edgeevidence}
\title{Edge evidence plot}
\usage{
plot_edgeevidence(output, evidence_thresh = 10, layout = "spring", ...)
}
\arguments{
\item{output}{Output object from the extract_results function}

\item{evidence_thresh}{BF which will be considered sufficient evidence for in-/exclusion}

\item{layout}{Layout of the network; qgraph argument}

\item{...}{Additional qgraph arguments}

\item{edge.width}{Layout of the network; qgraph argument}
}
\description{
Edge evidence plot
}
